requirements.txt
deepmerge
pexpect
gnupg
ovh
PyYAML
